+++ lib/rspec/wait.rb
@@ -7,12 +7,8 @@ require "rspec/wait/target"
 module RSpec
   module Wait
     # From: https://github.com/rspec/rspec-expectations/blob/v3.0.0/lib/rspec/expectations/syntax.rb#L72-L74
-    def wait_for(*args, &block)
-      if args.any? || block.nil?
-        raise ArgumentError, "You must pass only a block to `wait_for`."
-      end
-
-      Target.new(block)
+    def wait_for(value = Target::UndefinedValue, &block)
+      Target.for(value, block)
     end
   end
 end