+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.0.2 / Unreleased
 
-* [ENHANCEMENT] Allow `wait_for` to accept only a block target
+* [ENHANCEMENT] Allow `wait_for` to accept either a value or block target
 * [BUGFIX] Ensure blocks are re-evaluated with each iteration
 
 ## 0.0.1 / 2014-04-19